# [REF] account_check_printing: Speed-up payment creation