@moylop260 2022-01-15
Hi @nim-odoo @mart-e 

Could you check this refactoring, please?

It is 18k x faster if there are check_number
and reduce the number of times that the constraint is executed if there are not check_number