@fla-subteno-it 2022-01-25
ping @oco-odoo